+++ src/main/java/com/simibubi/create/content/contraptions/fluids/actors/SpoutTileEntity.java
@@ -108,7 +108,7 @@ public class SpoutTileEntity extends SmartTileEntity implements IHaveGoggleInfor
 		if (tank.isEmpty())
 			return HOLD;
 		FluidStack fluid = getCurrentFluidInTank();
-		int requiredAmountForItem = FillingBySpout.getRequiredAmountForItem(level, transported.stack, fluid.copy());
+		long requiredAmountForItem = FillingBySpout.getRequiredAmountForItem(level, transported.stack, fluid.copy());
 		if (requiredAmountForItem == -1)
 			return PASS;
 		if (requiredAmountForItem > fluid.getAmount())
